requirements.txt
requests==2.32.3
mysql-connector-python==9.0.0
pandas==2.0.3
numpy==1.24.4
Flask==3.0.2
Flask-Cors==5.0.0
uWSGI==2.0.23
python-dateutil==2.9.0.post0
pytz==2025.1
tzdata==2025.2
nltk==3.9.1
ftfy==6.3.1
regex==2024.11.6
scikit-learn==1.3.2
scipy==1.14.1
joblib==1.4.2
matplotlib==3.7.5
seaborn==0.13.2
python-dotenv==1.0.1
Telethon==1.39.0
uWSGI==2.0.23
